Renamed variable for clarity and updated references
The variable `array`, used to store the list of files in the `pictosFolder` directory, has been renamed to `pictoImageFiles` for better clarity. The `Console.WriteLine` statement and the `Parallel.For` loop have been updated to use this new variable name. These changes improve the readability of the code by using a more descriptive variable name.