Clarify error causes and exFat risks in documentation
Added details that error Panda-0008_0101 can be due to file not in cache or too many cache folders, with a limit around SD_Cache.0028. Updated exFat section to note removal of 4GB size limit but highlighted risk of data loss. Included a disclaimer about non-responsibility for data loss when switching to exFat.