Add "Speculation" section for error Panda-0008_0101
Added a new section titled "Speculation" under the error code Panda-0008_0101 in the ERRORS.md file. This section explains a potential cause for the error related to cache size issues and provides solutions such as reducing the number of caches or switching to the exFAT file system. A warning about the risks of using exFAT, including potential file or drive corruption, is also included, along with a disclaimer about data loss.