Add debug rethrow for exceptions in ConverterDialogue
This change introduces a conditional compilation directive that rethrows exceptions in debug mode within the `ConverterDialogue` class. This enhancement allows developers to debug exceptions more effectively when running the application in a debug environment. The rethrow is wrapped in `#if DEBUG` and `#endif` preprocessor directives to ensure it only occurs in debug builds.